+++ tests/ByteDev.Reflection.UnitTests/TestTypes/TestClasses.cs
@@ -45,4 +45,11 @@
             Value = value;
         }
     }
+
+    public class TestPerson
+    {
+        public string Name { get; set; }
+
+        public int Age { get; set; }
+    }
 }